+++ layouts/default.vue
@@ -1,10 +1,5 @@
-<script setup lang="ts">
-import KodeeWrapper from '../components/KodeeWrapper.vue'
-</script>
-
 <template>
   <div class="slidev-layout default">
     <slot />
-    <KodeeWrapper />
   </div>
 </template>